+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Change Log
 0.0.18
 -------
 - Support ROE %[`#8 <https://github.com/defeat-beta/defeatbeta-api/issues/8>`_]
+- Support ROA %[`#51 <https://github.com/defeat-beta/defeatbeta-api/issues/51>`_]
 
 0.0.17
 -------